test(fip0118): split solstice itests to fit CI, fix lint/gen, changelog
Split the oversized NV29 solstice itest files so each _test.go stays
under the 10m CI budget, ensure every split file compiles standalone,
and fix errcheck/context lint findings plus the generated CLI docs and
CHANGELOG entry, part of #13723
