+++ itests/solstice_miner_test.go
@@ -385,12 +385,12 @@ func TestMigrationNV29SolsticeAccounting(t *testing.T) {
 	// its 10x QAP once the miner's deferred termination cron fires at the end of the proving period.
 	um.TerminateSectors([]abi.SectorNumber{legs[1]})
 	expectedQA := uint64(defaultSectorSize) * (1 + 10) // remaining: legs[0]=1x + legs[2]=10x
-	solsticekit.WaitForMinerQAP(t, ctx, client, maddr, expectedQA, 2*time.Minute)
+	solsticekit.WaitForMinerQAP(ctx, t, client, maddr, expectedQA, 2*time.Minute)
 
 	// ---- Terminate x CC-1x: terminating a never-upgraded legacy CC sector (1x) must drop QAP by
 	// exactly its 1x contribution once the deferred termination cron fires, leaving only legs[2]=10x.
 	um.TerminateSectors([]abi.SectorNumber{legs[0]})
-	solsticekit.WaitForMinerQAP(t, ctx, client, maddr, uint64(defaultSectorSize)*10, 2*time.Minute)
+	solsticekit.WaitForMinerQAP(ctx, t, client, maddr, uint64(defaultSectorSize)*10, 2*time.Minute)
 }
 
 // TestMigrationNV29SolsticeEconomic runs on an unmanaged miner with identical legacy CC sectors so
